requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
wandb>=0.16.0
numpy>=1.24.0
pytest>=7.0.0
jupyter>=1.0.0
matplotlib>=3.7.0
pandas>=2.0.0
anthropic>=0.7.0
weave>=0.20.0